+++ src/modules/map/store/MeasurementObserver.js
