+++ test/modules/map/components/olMap/handler/measure/OlMeasurementHandler.test.js
@@ -13,7 +13,7 @@ import { $injector } from '../../../../../../../src/injection';
 import { TestUtils } from '../../../../../../test-utils.js';
 import proj4 from 'proj4';
 import { register } from 'ol/proj/proj4';
-import { MEASUREMENT_LAYER_ID } from '../../../../../../../src/modules/map/store/measurement.observer';
+import { MEASUREMENT_LAYER_ID } from '../../../../../../../src/modules/map/store/MeasurementObserver';
 
 
 const environmentServiceMock = { isTouch: () => false };
